+++ src/main/resources/schema.postgresql.sql
@@ -47,8 +47,8 @@ alter table fitbit_heartrate_
 create table fitbit_cardioscore_ (
     user_id_                varchar(32) not null,
     date_                   date not null,
-    score_min_              integer null,
-    score_max_              integer null,
+    score_min_              double precision null,
+    score_max_              double precision null,
     score_estimated_        double precision null
 );
